backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
anthropic==0.28.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1
pydantic==2.12.5
prometheus-fastapi-instrumentator==6.1.0
python-json-logger==2.0.7
httpx==0.27.0
